Use InstanceType<typeof> instead of explicit interfaces
